[REFACTOR] Refactored all of Infyom auto-generated files from Events to Event to stick with Laravel naming conventions
At the time the command was run, some models were written with the wrong naming convention. To avoid loosing work, the "easiest" way of refactoring this was file by file, line by line. Please use this commit for further refactors and if something goes wrong, this commit is here to help find the problems

Signed-off-by: Daniel Graça <danielgrascas@gmail.com>