Changed way venues list are displayed to user
Changed the way a venue is listed to the user in the venue's dropdown. Now it uses the format: name (address) where the address is stripped to a max of 86 char. Why? Well, the maximum number of characters possible to be stored in the DB is 512. So why not? Could and probably will be changed in the future.

Signed-off-by: Daniel Graça <danielgrascas@gmail.com>