[IMPORTANT] Removed option to associate Event with Entity. This commit has comment explaining why.
The reason why the select box for associating Event with Entity was removed is simply because everywhere else it's being done automatically to user's default entity.

The form is ready to implement it in the future, where the User will be able to select which entity to use, not only here but as well when creating a Venue (form not ready yet).

Why should the user be able to choose? In the future, it could - and probably will - be good to have collaborators in an entity helping the event's manager... well managing the event.

But why? Why not just do it now? Do it while I can, later I'll have to update code, be angry and complain about it. Yikes! I'm feeling the pain already! But I think I can explain it. The reason is: prototype. I'm way happier to have a prototype done in time and working and then spend time tweaking the features than spengind time to have every single feature and OOPS time's out. I can't forget that I'm still learning a stack that is somewhat new to me (TALL) and due to the projects' complexity I would rather move on. Let's just not think that I spent an eternity writing this. Hmkay? Aight. Cheers.

Signed-off-by: Daniel Graça <danielgrascas@gmail.com>