Creation of event form WITHOUT validations. See comments
Allows to create an Event, with its sessions and tickets. All associations working.
[DISCLAIMER] Once again, no validation is being done. So, if an event is created without a Venue associated, it will crash and return an error.

It is possible to list all Events in {url}/admin/events, and even delete events selecting them in this page.

The show/edit event function needs to be worked on, because:
- It's possible to enter event details page, but no detail will be shown (and lacks Tickets and Sessions details) and when editing the form needs to be tailored.

On a final note: Breadcrumbs and routes were edited, as well as some props variables being injected in some Views beucase of the Infyom script (some variables were plural, it needed to be singular, as well as the breadcrumbs).

That's all for now. If you read this, hats off to you. Cheers.

Signed-off-by: Daniel Graça <danielgrascas@gmail.com>