[Tweak] Create event form allow association between one ticket and multiple sessions (1-n instead of 1-1)
The form had a dropdown in the tickets area to associate the ticket with a session. How it works with checkboxes. It is working both on create and update.

In the view, when creating/updating an event, it won't let delete tickets nor sessions if the number of tickets/sessions is 1. To prevent an empty form.

Also the eloquent relationships were changed.

The form needs validation, CSS tweaking (that :tw-merge type of behavior) and change some fields to use noop's Country and Country number selectors.

Signed-off-by: Daniel Graça <danielgrascas@gmail.com>