[README] Generate PDF and send it via email asynchronously
Refactored AccessTicketController.php so each method has its purpose.

GenerateAccessTicketPDF.php is responsible to generate the PDF

SendAccessTicketEmail.php is responsible to send the Email (soon will be a Notification instead of Mail)

The .docx file is test.docx, and it should be removed when the new one is provided by noop.

The command to run asynchronously the jobs in queue is *php artisan queue:work*

Signed-off-by: Daniel Graça <danielgrascas@gmail.com>