[FIX][BUG][IMPORTANT] Renamed a column in the original migration. PLEASE README
This shouldn't be done. EVER! Migrations' purpose is to change data, without breaking what's in production. With that said, this had to be done because a column name was badly structured. I don't know how the migrations were able to run. This fixes it and also fixes a bug where pre-approval wasn't being properly set in DB.