Ticket purchasing validations, making use of event state machine and adding code to help future ticket pdf tempate
When someone buys tickets, it validates if the email inserted has already surpassed the limit of tickets allowed per account.

Started using the Event state machine. When an event is in the state of draft, the form to buy tickets is unreachable. If the event is available, tickets may be purchased. If the event is closed, cancelled or finished, it is not possible to buy tickets.

Passed eventSessionTickets into the CustomPdf.php helper (and into the blade file to create the access tickets PDF) in order to have that information, so in the future it's possible to add that into the PDF template.