Added "paid" flag to access_tickets database structure
This flag will be used to determined if a ticket can be used or not. The ticket can/will be verified manually or automatically, depending on the payment method.

At this point I'm kind of lost, but I think I might implement 2 ways of paying:
- Bank transfer (per noop's request)
- MBWay or PayPal (per college requirements)

NOTE: MBWay is a portuguese service that can be used internationally IF the phone number is associated with a portuguese bank that subscribes to this service. Because of this, and because it would be nice this service to be available internationally, to have PayPal.