[README] Payment options are stored in specific key/value format
The payment options data is stored in json in the DB with key/value. This will be used later to define the multiple ways of paying for a ticket (instantly with PayPal or via bank transfer).

PayPal will automatically define the ticket as paid, whereas the bank transfer will force someone from the managing team to manually validate the ticket as paid.