[README][IMPORTANT] Multiple ways of paying (bank transfer and paypal)
This commit changes the way this web application deals with payments. Now the user can choose between bank transfer or PayPal if - and only if - the managing team fulfilled their information with their banking details (for bank transfer) and with their paypal email (for paypal). If they choose to use only PayPal, or only bank transfers, it's their decision.

What happens if they don't fill in their information?
Well... the user can see the page to buy tickets, but cannot buy them as there is no button to buy :) #wompWomp