[README] Added cron jobs to the project to delete unpaid access tickets and restore the tickets stock
This commit allows the server to continuously run cronjobs to delete unpaid access tickets and restore them to the stock.

While doing so, I've thought about it and it is extremely difficult to send email with a PDF with all the tickets bought at once when they are paid. This is because automatically (or manually) the tickets are being set as paid individually. So, in my humble opinion, I would say that the best approach would be to have a "transactions" table, where the whole transaction is set as paid instead of each access ticket. If this is confusing, let me explain below:

What is implemented now:
Lets say that I purchase 3 tickets for the event XPTO. This event has 2 sessions - A and B - and each session has 2 tickets - basic and premium.

Now lets say I bought 1 premium for session A and 1 premium and 1 basic for session B.

What the web application does in the background is "Okay. Someone bought this 3 tickets, lets create the respective access tickets and check if they are *individually* paid. If they are, good. If they aren't, delete them (just after X amount of hours)".

It would be better if the application would just "Okay, there's an X transaction that was made by someone, let's just see if the transaction was paid. If it wasn't, lets rollback and restore the stock. If it was, lets send an email with all the tickets from this transaction."

I think this might be a better approach.