Manually set as paid and approved is now in transactions
Before the managing team would have to manually mark each access ticket bought as paid or approved. If someone bought 10 tickets, the team would've to mark all 10 tickets as either paid or approved. This could lead to some trouble or even to forget to mark some tickets.

Now the transaction (which is what identifies ALL the tickets purchased, e.g. 10 tickets) is marked as payed. Now the user just needs to say what is the transaction ID or code (doesn't have any code, but could be a potential addition) and the team would just need to approve that transaction. If the user fails to pay, in 72h the transaction would be reverted and deleted.