[BUG FIX][TWEAK][BASE PROJECT FIX][README] Multiple project tweaks and project base changed
The int-tel-input phone package was changed the way that it loads in the view, allowing to use the phone component in modals and other views.

Added phone components to all components that needed the int-tel-input.

Fixed the bug when the cancelling a form would submit it as "save" and it wouldn't work.

Changed the way of validating a ticket (one less route).

Started working on the Home page (admin dashboard).